[accumulators] Move code from feature/binning_analysis.hpp to .cpp
I had to comment out converged_errors() method because it was broken and
did not compile. Apparently, this problem had not been detected before
because the method never got instantiated...
